backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
python-dotenv==1.0.0
mangum==0.17.0
